+++ src/test/java/fr/utc/miage/shares/ActionTest.java
@@ -138,7 +138,7 @@ class ActionTest {
          * @return always 0
          */
         @Override
-        public float valeur(final Jour aJour) {
+        public double valeur(final Jour aJour) {
             return 0.0F;
         }
     }
